requirements.txt
langchain-chroma==0.1.4
langchain_ollama==0.2.0
langchain-huggingface==0.1.2
chromadb==0.5.18
fastapi==0.115.0
python-dotenv==1.0.1
langserve==0.3.0
uvicorn==0.31.0
sse-starlette==2.1.3
langchain_community==0.3.5
pymupdf==1.24.13
nltk==3.9.1
opentelemetry-api==1.27.0